refactor: ProfileLarge to molecule
